+++ project/agenttrainer.h
@@ -9,8 +9,8 @@ class AgentTrainer
 public:
     AgentTrainer();
     void train(A* agent, int numberOfEpisodes,int trainMode = 0,int savingSequenceMode = 0);
-    void saveEpisode(vector<vector<double>> stateSequence, int seqId);
-    vector<vector<double>> loadEpisode(int seqId);
+    void saveEpisode(vector<vector<float>> stateSequence, int seqId);
+    vector<vector<float>> loadEpisode(int seqId);
 
 };
