+++ project/discreteaction.cpp
@@ -10,7 +10,7 @@ DiscreteAction::DiscreteAction(int size): size(size)
 
 }
 
-double DiscreteAction::pick()
+float DiscreteAction::pick()
 {
     default_random_engine generator(random_device{}());
     uniform_int_distribution<int> dist(0,size-1);
